requirements.txt
Django==1.9.8
djangorestframework==3.4.0
